Update root disk detection in is-rootdisk-device.sh
The root filesystem is now carved out of the cgts-vg volume group
instead of a dedicated partition. Update the is-rootdisk-device.sh to
look for '/boot' instead of '/' as '/boot' is still a dedicated
partition on the root disk.

This will impact proper I/O scheduler selection.

Test Plan:
PASS - Build/Install. Execute script passing root an non-root disks.
       Make sure proper output is displayed

Signed-off-by: Robert Church <robert.church@windriver.com>
Story: 2010444
Task: 47000
Change-Id: I77491e02adb51bc85ea7c3b3a1b89eda7c91a1a5