Generate checksum across ostree_repo for prestage validation
This commit calculates an md5-based checksum across the ostree_repo
directory. The checksum is included in the ISO, available for the
kickstart files to validate the integrity of the ostree_repo before it
is used as the primary boot repo for the subcloud. In the remote install
case, we use this checksum to invalidate prestage data if necessary, in
which case it falls back to performing a remote pull from the system
controller.

Test Plan
PASS:
- Run gen-prestaged-iso.sh to generate a bootimage.iso. Verify that
  the ISO includes a .ostree_repo_checksum file at the same level as the
  ostree_repo directory.

Partial-Bug: 1999306

Change-Id: I054ef69a243cb26d5c8a4cb0ce74e289b63670c1
Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>