Setting the default time out for prestaged iso install
When we install the prestaged iso, the first screen we are
presented with is the grub menu screen on the subcloud. This
should timeout after a default 30 seconds, and then start the
installation procedure.

This commit ensures that the installation begins by default after
30s.

Test Plan:

PASS: Verify that the first screen times out after 30 seconds
      and boots into the highlighted option.

PASS: Verify that the screen times out after the time configured
      with the -t option when creating the prestage iso.

Closes-Bug: 1999565

Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>
Change-Id: I8a1ea19cebd2f1637e5626f2438220247edb5711