Warn about invalid date when collecting logs
If the user entered an invalid start-date, invalid end-date and/or
the start-date is greater than the end-date during the collect logs
process, the system will warn the user about the error, instead of
just reverting to collect all.

Test plan:
PASS - Install and bootstrap SX and test if the warnings are
presented if an invalid start-date, end-date and/or date range
are passed by the user, as well as collecting logs if correct
parameters are provided.
PASS - Install and bootstrap Standard and test if the warnings are
presented if an invalid start-date, end-date and/or date range
are passed by the user, as well as collecting logs if correct
parameters are provided.

Closes-bug: 2003791

Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>
Change-Id: Iab35b1d92f4471e2713fabe5d7a6ef8b611ea240