Warn about invalid date when collecting logs
If the user entered an invalid start-date, invalid end-date and/or
the start-date is greater than the end-date during the collect logs
process, the system will warn the user about the error, instead of
just reverting to collect all.

Test plan:
PASS - Install and bootstrap SX and test if the warnings are
presented if an invalid start-date, end-date and/or date range
are passed by the user, as well as collecting logs if correct
parameters are provided.
PASS - Install and bootstrap Standard and test if the warnings are
presented if an invalid start-date, end-date and/or date range
are passed by the user, as well as collecting logs if correct
parameters are provided.

Test with correct parameters:
PASS: collect --start-date any --end-date any
PASS: collect
PASS: collect --all
PASS: collect --list controller-0 controller-1
PASS: collect controller-0 controller-1
PASS: collect --all --subcloud

Closes-bug: 2003791

Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>
Change-Id: Iab872d2cbf02e1e6b6feef931375d05345c844e8