Update debian package versions to use git commits
The Debian packaging has been changed to reflect all the
git commits under the directory, and not just the commits
to the metadata folder.

Package builder concatenates the number of revisions added since
the one defined in the BASE_SRCREV to the name of the .deb file
generated. Thus, any new submission to the code in these directories
increments this revision counter in the filename.

Test Plan:
PASS: Created new code submissions to package python-cephclient
and verify that the version in the package name was incremented.
PASS: Created new code submissions to package ceph-manager and
verify that the version in the package name was incremented.

Story: 2010550
Task: 47510

Signed-by: David Bastos <david.barbosabastos@windriver.com>
Change-Id: I365f01789bb129f869ce52a3372bb99bbc5ab548