Disable parallel collect mode when there is only one host
CPU consumption is increased when using the parallel mode in
comparison to inline mode, due to the task monitoring code, even when
there is just one host to collect.

To improve collect performance on AIO-SX system configured with a
single platform core, this change disables parallel collection mode
when there is just one host and it is not a subcloud collect.

TEST PLAN:
PASS: AIO-SX: Perform a collect without this change and notice the cpu
      consumption on /var/log/collectd.log (it is shown as ssh
      process). Manually edit the source code with this change and
      perform the collect again. Notice the cpu consumption is
      decreased.
PASS: STD: Perform the same test from AIO-SX on active controller node
      but specify the inactive controller for collection.
PASS: STD: Perform the same test from AIO-SX, but from inactive
      controller.

Story: 2010533
Task: 47600

Signed-off-by: Alyson Deives Pereira <alyson.deivespereira@windriver.com>
Change-Id: Ide743d33ad1e0af40344fd4faa8153388891c655