Add initial password option to update-iso.sh
Sometimes, for security purposes, the user might want to change on the
ISO the initial password prompted on first login, that is "sysadmin"
by default. This change adds the --initial-password option to
address this. The user will still be prompted to change the password
on first login regardless.

Test plan:
  PASS: Run update-iso.sh with --initial-password, install the new ISO
        and verify that the provided password is the initial password.
  PASS: Run the same test with different ISO builds.

Closes-Bug: 2013365

Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>
Change-Id: I267968f322771a0a3d17c0ad7742ebd401d9b08e