Update gen iso scripts for CentOS support in Debian
Update existing dynamic ISO generation scripts to target the last
CentOS release from within a Debian environment.

Changes are applied for gen-prestaged-iso-centos.sh for support
within a Debian instance.

The existing Debian-targeting files (gen-bootloader-iso.sh,
gen-prestaged-iso.sh) are updated for minor fixes in logging,
consistency, and overall code structure for ease of support and
maintenance.

This commit aligns recent changes of gen-bootloader-iso.sh into
gen-bootloader-iso-centos.sh, for easier support going forward.

And finally, unit tests are added for the gen-prestaged-iso* scripts.
The unit tests use shunit2 framework, which is dynamically pulled in as
required to run the tests. At this point, tests are only run manually;
they are not part of the loadbuild. Unit tests may be added for
gen-bootloader-iso.sh in a further commit; however, this is of lesser
priority since this script is employed frequently on subcloud remote
installations.

Test Plan

PASS:
- Test gen-prestaged-iso.sh for generating a prestaged ISO targeting a
  local Debian install. An ISO is generated and a full subcloud
  deployment is successful
- Test gen-prestaged-iso-centos.sh for generating a prestaged ISO
  targeting a local CentOS install. An ISO is generated and a full
  subcloud deployment is successful
- Run the two new unit tests, gen-prestaged-iso-test.sh, and
  gen-prestaged-iso-centos-test.sh, both in standalone and via
  run-tests.sh. Verify with both an empty and populated input directory.
- Verify full remote/redfish Debian subcloud install and deployment.
  Targeted at gen-bootloader-iso.sh changes.
- Verify full remote/redfish CentOS subcloud install and deployment.
  Targeted at gen-bootloader-iso-centos.sh changes.

Story: 2010611
Task: 48267
Depends-On: https://review.opendev.org/c/starlingx/metal/+/886662

Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Change-Id: I31e76ed107f589b1196320b3c8d7243fb15d3491