Fix openssl passwd args for initial password change
This commit fixes the incompatible --crypt option for older
releases of openssl. We also add the -quiet option to suppress
the warning message if a password is longer than 8 characters.

Test Plan:
PASS: Run this script with --initial-password option in centos
      environment where openssl version is:
      OpenSSL 1.0.2k-fips  26 Jan 2017
PASS: Run this script with --initial-password option in debian
      environment where openssl version is:
      OpenSSL 1.1.1n  15 Mar 2022

Closes-Bug: 2028787

Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Change-Id: I2e87a6e8c1fa9fe4d3c1dc8982995a09f997f47b