update-iso.sh: add --no-force-password option
The --no-force-password option removes the forced password change on
initial login. This is to enable customers who use a customized workflow
where the initial password is set to a secure value via update-iso.sh,
and do not want a forced password change on initial login.

Note: Used incorrectly, this is an insecure option. The
--no-force-password should only be used in secure environments, where
the initial password is kept secret. Further, this update does not
affect the password aging policies in place.

Test Plan:
PASS: Use update-iso.sh to change an existing ISO:
    sudo ./update-iso.sh --addon ./ks-addon.cfg \
      --no-force-password \
      --initial-password 'MyPassword#123' \
      --input ./starlingx-intel-x86-64-cd.iso \
      --output ./update.iso
And boot from the new ISO. Verify that the initial password upon
boot is as expected, and that the provide password is accepted on first
login, without the normal forced password change on initial login.

PASS: Use update-iso.sh without the --no-force-password option, verify
that the user is still forced to change password on initial login.

Closes-Bug: 2030917

Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Change-Id: I962e66741cf10b24ba965bef99f3cd2ef977cec4