Add missing dependencies in stx-ceph-manager debian image
stx-ceph-manager images are actually being generated for both
CentOS and Debian images. When comparing both images, we can
notice that Debian images are missing the following dependencies:
python-cephclient and ceph-manager. Therefore, rook-ceph cannot
be applied.

To resolve this, dependencies were added to the Dockerfile of
the stx-ceph-manager image.

Test Plan:
 PASS: Build stx-ceph-manager image
 PASS: Push image to local registry
 PASS: Apply rook-ceph-apps
 PASS: Check if ceph is healthy

Closes-Bug: 2039564

Change-Id: Ie03cbead00f26bf972c7e4796d9f8472b58b91c9
Signed-off-by: Erickson Silva <Erickson.SilvadeOliveira@windriver.com>