Change CPU affinity script to help reduce upgrade activation failures
Currently the script (affine-tasks.sh) waits until critical platform
pods are recovered and will then affine all platform tasks back to
the provisioned platform cpus.  This leads to periods of time of 1-2
minutes during upgrade activation where platform cpus are running
close to 100% and this contributes to rabbitmq problems.

This change is to affine platform tasks only when platform upgrade
is activation-complete on a active-controller.

This removes workerconfig.service dependency as it is not present on
AIO and the start() of the script exits if the configuration is not AIO.

Test Plan:
PASS: Perform platform upgrade stx7 to stx8 in AIO-SX, AIO-DX and
      confirm that all platform tasks are affined to platform cores
      after the step activation-complete
PASS: Performed complete platform upgrade, sucsessfully unlocked
      the node.
PASS: Install AIO-DX + worker, confirm Bootstrap completes and
      standalone worker node is unlocked.
PASS: Built package successfully.

Closes-Bug: 2038537

Signed-off-by: Jim Gauld <james.gauld@windriver.com>
Change-Id: Ifc863587f8ffaf6b6478ae9edc352449430824b7
Co-authored-by: Saba Touheed Mujawar <sabatouheed.mujawar@windriver.com>
Co-authored-by: Jim Gauld <james.gauld@windriver.com>