Improve logs and case-sensitive on k8s-coredump
Improving logging by adding timestamp and
show error in case we have a KeyError while
trying to verify configurations set by
pod annotations.
Also allowing case-sensitive paths to be
configured as the path to save the coredumps.

PASS: AIO-SX install and bootstrap.
PASS: On a AIO-SX test if k8s-coredump is executed
	successfully.
PASS: On a AIO-DX test if k8s-coredump is executed
        successfully.
PASS: On a Standard test if k8s-coredump is executed
        successfully.
PASS: On a Standard test if k8s-coredump is executed
	successfully using a PVC setup.

Partial-Bug: 2045789

Change-Id: I942253468a596b7d68d6a3733b24062cfb4661f7
Signed-off-by: Heron Vieira <heron.vieira@windriver.com>