Fail fast improvement to show-certs
When k8s certificates are expired, this script can take more than 1
minute to return. During debugging, I noticed that the most time
consuming step is PrintCertInfo-for-OIDC-Certificates, which queries
for kubernetes secret many times. PrintCertInfo-fromGenericSecret,
below it, also makes a few calls with kubectl.

This commit adds a variable to capture the return of 'kubeadm certs
check-expiration' command and then only calls
PrintCertInfo-for-OIDC-Certificates if the RC is successful.

This reduces the overall execution time from about 1 minute to around
10 seconds.

Test Plan:
PASS: Run show-certs in c0 and c1 and verify it finishes successfully
      and the output before and after this change is the same
PASS: Cause k8s certificates for expire. run show-certs in c0 and c1 and
      verify it finishes successfully in less than 10s. Verify that the
      output before and after this change is the same.

Story: 2010815
Task: 49485
Change-Id: I9f3eaec3a543fdea278e04c2f1895685bc333505
Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>