Add cloud-init boot menu option to gen-prestaged-iso.sh
This commit introduces a structured main menu and sub-menus to
the boot menu, accommodating new cloud-init installation requirements.
The new boot menu is organized as follows:
  - "Prestage Install"
  - "Prestage cloud-init All-in-one Install"
  - "Prestage cloud-init All-in-one (lowlatency) Install"

Each sub-menu offers Serial and Graphical Console options to cater to
different installation preferences.

Cloud-init boot menu options are added to enable the cloud-init service
on the boot of a prestaged ISO, allowing it to be booted with a
NoCloud seed datasource, which is used in the factory install workflow.

Additional fix: when a ks-addon kickstart script is added, it renamed
to ks-addon.cfg. This is because the main kickstart.cfg only looks for a
file named ks-addon.cfg; if we don't rename the file it will be ignored.

Minor fix: add colour to highlight script failures and warnings.

Test Plan:
1. PASS: Genereate prestage ISO. Boot and verify that selecting
         cloud-init is working as desired. The cloud-init services
         are active. Boot with a seed ISO; verify that cloud-init
         is properly executed. Verify with the various installation
         configurations, both graphical and serial for:
          - "Prestage cloud-init All-in-one Install"
          - "Prestage cloud-init All-in-one (lowlatency) Install"

2. PASS: Verify boot args for each selection

3. PASS: Ensure default-boot parameter correctly selects
         the desired boot option.

3. PASS: Verify ks-addon.cfg renaming inside the prestaged ISO

Story: 2011100
Task: 49970

Depends-on: https://review.opendev.org/c/starlingx/metal/+/914995

Co-authored-by: Salman Rana <salman.rana@windriver.com>
Change-Id: I174678e768b3f040a5cd07a8d643a3d825a55442
Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Signed-off-by: Salman Rana <salman.rana@windriver.com>