Fix utf-8 decode traceback in the report tool strings algorithm
This update adds a try or except on UnicodeDecodeError to the
utf-8 decode case in the _evaluate_substring function of the
strings algorithm.

Added a dropped logs file to record a list of logs that fail the utf-8
conversion. The dropped logs file is stored as dropped_logs in the
report analysis plugins output folder.

Test Plan:

PASS: Verify on collect bundle that exhibited the issue.
PASS: Verify dropped logs are recorded in new
      report_analysis/plugins/dropped_logs file.
PASS: Verify no newly introduced pep8 errors or warnings.

Closes-Bug: 2067952
Change-Id: I82cecd6bb537a459255467f6ced7e742c5771d97
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>