stx-debian-{tools-}dev: add build-essential
A recent review [1] removes build-essential from the base image,
stx-debian. Re-add it here, so that we don't depend on build-essential
being part of the base image.

NOTE: this package is not in the managed debian repo created by the
download lists in stx-tools, so we install it from debian snapshots, the
same way the base image did prior to [1]. We plan to add these packages
to the download lists in a separate commit.

[1] https://review.opendev.org/c/starlingx/root/+/922463

TESTS
===========================
* Build stx-debian from [1]
* Build stx-debian-dev & stx-debian-tools-dev

Story: 2011159
Task: 50405

Change-Id: Ifc38620cb49416aebcff59b9926af4c60a1005e4
Signed-off-by: Davlet Panech <davlet.panech@windriver.com>