Fix collect_host exclude list handling
The Collect Tool has a feature that omits a list of files
or directories that should be excluded from the collect bundle.
This is referred to as an exclude list.
There is a similar include list.

Collect is not always honoring what is specified in the exclude list.

Both the include and exclude lists are passed to the 'tar' utility.
Tar prioritizes the include list contents over the exclude list
contents for any file or directory that exists in both lists.
This is not correct behavior for collect.
The exclude list should take priority.

This update introduces a filter function that ensures any dir or file
in the exclude list is removed from the include list and then passes
the potentially updated include list to the tar utility.

Test Plan:

PASS: Verify build install and 'collect all' in a AIO DX system.
PASS: Verify the contents of the exclude list is prioritized
      over the include list when there is the same entry in both lists.
PASS: Verify a file that is specified in the exclude list
      is not collected.
PASS: Verify a directory that is specified in the exclude list
      is not collected.
PASS: Verify handling when there is no duplication between the lists.

Closes-Bug: 2070671
Change-Id: Ifc117803b4d675d137e1d807652e90ccf9212191
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>