Get certificate with mgr name instead of hostname
In bare-metal ceph, mgr's were created with the hostname,
however with rook-ceph they are created with letters.
Therefore, it is not possible to get the certificate
from restful API module because the hostname is passed
instead of the mgr name.

To fix this, the "ceph mgr metadata" command has been
replaced with "ceph mgr dump", which returns the name
of the active mgr.

Test Plan:
- PASS: Build stx-ceph-manager image
- PASS: Change the stx-ceph-manager deployment in
	the rook-ceph app to use this image
- PASS: Build rook-ceph app
- PASS: Apply app on Standard (IPV6)
- PASS: Check the pod's logs to verify that
	the certificate was obtained
- PASS: Using bare-metal ceph, check the logs in
        /var/log/ceph-manager.log for any errors.

Story: 2011066
Task: 50656

Change-Id: I4f57e1d0cc45143c1ea29c826e45bae59e27efdc
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>