Change stx-ceph-manager base image
The stx-ceph-manager image, which is used for rook-ceph,
was with ceph v14.2.2, which is the bare-metal version.
However, rook-ceph uses ceph v18.2.2.

Therefore, to maintain version compatibility, the base image
was updated to docker.io/openstackhelm/ceph-config-helper:
ubuntu_jammy_18.2.2-1-20240312 and all necessary adjustments
were made to the Dockerfile so that all STX packages could be
installed successfully.

Test Plan:
- PASS: Build stx-ceph-manager image
- PASS: Change the stx-ceph-manager deployment in
	the rook-ceph app to use this image
- PASS: Access the stx-ceph-manager pod and check
	the ceph version
- PASS: Check the pod logs to verify if there
	are any errors

Story: 2011066
Task: 50702

Depends-On: https://review.opendev.org/c/starlingx/utilities/+/924883
Depends-On: https://review.opendev.org/c/starlingx/app-rook-ceph/+/924884

Change-Id: I30371e7ed586f5b956668517bc1f9e5ed6b5cdce
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>