Add back the install types 4 and 5 for backward compatibility
This commit adds back the install types 4 and 5 (lowlatency). It
allows the operator install a subcloud N-1 as rt-kernel and
lowlatency in the subfunctions on a system controller running
the current release on system controller.

Test Plan:
PASS: Install a N-1 subcloud as rt-kernel and lowlatency in the
subfunctions on a current software release on the system controller.

Failure Path:
PASS: Ensure that the script does not accept the install types 4 and 5
on current software release and return an error.

Partial-Bug: 2084830
Change-Id: I1a2285c9867a4c040306a235b848181e2930976f
Signed-off-by: Wallysson Silva <wallysson.silva@windriver.com>