script to ensure that AppArmor is enabled
AppArmor is disabled on hosts by default. This change introduces
the `host-cis-benchmark-apparmor-setup.sh` script, which can be
used to enable AppArmor on a host if it is not already enabled.
The script checks the current AppArmor status and only performs
changes if necessary.

Test Plan:
PASS: Build an ISO and deploy an AIO-SX environment.
PASS: Verify that `host-cis-benchmark-apparmor-setup.sh` exists
      at `/usr/local/bin` in the AIO-SX setup.
PASS: Execute `host-cis-benchmark-apparmor-setup.sh` and verify
      that AppArmor status changes to 'enabled'.
PASS: Run `aa-status` to confirm AppArmor is enabled after the
      script runs and host is unlocked.
PASS: Execute the script when AppArmor is already enabled on the
      host and confirm that the script exits gracefully without
      changes.

Story: 2011253
Task: 51257

Change-Id: Ia96bd1203c7e47ee29292b0328dfc25a888f49cb
Signed-off-by: Jagatguru Prasad Mishra <jagatguruprasad.mishra@windriver.com>