K8s configurations for apiserver CIS compliance
This script handles the CIS errors related to the apiserver raised
by the KUBESCAPE tool. This script ensures k8s configs
are applied properly to the apiserver and it also ensures
sanity of the api-server after configs are applied.

Tests
PASS: Check api-server is up and running before and after executing script.
PASS: Check plugins applied successfully in the api-server.

Story: 2011337
Task: 51634

Change-Id: I2038e323c4bcda1bdab996e646ec1dab9c322ecd
Signed-off-by: Sabyasachi Nayak <sabyasachi.nayak@windriver.com>