Change final collect bundle to remove world read permission.
"collect" generates a bundle file with read permissions for
"other". This allows unauthorized users to view confidential
information within the generated tarball.

The fix changes the permissions of the bundle file right after it
is created.

Confirming Presence of Fix:

To confirm the presence of this fix in an installation run the
following command. If the result is 0 the fix is not present:

sysadmin@controller-0:/scratch$ grep -c chmod /usr/local/bin/collect
1
sysadmin@controller-0:/scratch$

Test Plan:

PASS:- Test 1 - All Hosts: run the default "collect" and confirm the
                tarball permissions.
PASS:- Test 2 - Individual Hosts: run the "collect --list" command
                where at least one host is remote (e.g. standy
                controller) and confirm the tarball permissions.

Regression Tests:

Due to the limited nature of the fix only the sanity tests listed above
are needed.

Note - the Starling X Bug listed below is a valid bug. However,
       it is hidden because it is a security bug.

Closes-Bug: https://bugs.launchpad.net/starlingx/+bug/2105451
Change-Id: I2fe4f27bb5281bdcad25c991322e8963a61b42b4
Signed-off-by: Tim Rose <kevin.rose@windriver.com>