Generate VLAN configuration via cloud-init without rebooting the host
The current implementation for configuring the OAM interface
during enrollment does not create the VLAN configuration,
which prevents the OAM from being accessible.

This commit allows the VLAN configuration to be created via
cloud-init, without the need to reboot the host, allowing
the OAM interface to remain accessible.

Test Plan:

PASS: - Bring up a subcloud with factory-install and verify
        that there is no access through the OAM. The VLAN
        will be configured when enrollment is performed.
      - Run the enrollment subcloud through dcmanager and
        verify that it completes successfully. The VLAN
        will be configured once enroll-init completes.
      - The enrollment should complete without the need
        to reboot the subcloud.

Closes-bug: 2109795

Change-Id: I761db307f2e05a34082a746875aa72510355f357
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>