Force ifup during cloud init
During enroll init process, if the OAM address is on the same device,
but assigned a different address, the ifup command is paused silently
before creating the new address (the return code is still 0). This
behavior can cause the default route to be lost, with subsequent
new default route creation failure.

This commit adds the force option to enforce the ifup finish its
tasks, ensures new IP address created on the interface. This commit
also adds additional logs for debugging because of the behavior of the
ifup command.

Test plan:
1. Passed - Enrollment test on subcloud with same OAM VLAN but different OAM IP
2. Passed - Enrollment test on subcloud with different OAM VLAN and IP
3. Passed - Enrollment test on subcloud with same OAM IP and VLAN.

Closes-bug: 2109795
Change-Id: I662721d3e1e81af0324fe4fc83ca352425f497e4
Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>