Limiting parallel execution of collect script
To limit the parallel execution of collect script, this commit intro-
duces a logic to validate if the node have an already running collect
instance and blocks new command calls, notifying the user.
This change affects all cluster configurations (SX, DX, STD, and DC).

In cases of attempted parallel execution of 'collect,' the standard
behavior is to notify the user and exit.

TEST PLAN:
PASS: Successfully execution of the collect command.
PASS: Successfully blocks new instances and notifies the user if collect
      is started while another already is running on the same node.
PASS: Successfully identifies and notifies the user if a new instance is
      started on some node (like stand-by controller or compute)
      while another instance is running in active controller (with --all
      option).
PASS: Successfully identifies and notifies the user if a new instance is
      started on Central Cloud active controller (with -sc option)
      while another instance is running on the subcloud.
PASS: Successfully identifies and notifies the user if a new instance is
      started on Central Cloud active controller (with -sc option)
      while another instance is running on some node (like stand-by
      controller or compute) on the subcloud.

Story: 2011322
Task: 52215

Change-Id: I6493947fb5c701667c641953879cc8e406faa2a9
Signed-off-by: Eduardo Alberti <eduardo.alberti@windriver.com>