Skip the factory backup stage on duplex
This script is a follow-up to [1], aiming to skip the execution
of the backup script during the factory backup stage on duplex
systems.

With this change, the script detects if the system is duplex, logs a
message, and returns early, ensuring that the factory install is not
blocked.

[1]: https://review.opendev.org/c/starlingx/utilities/+/954378

Test Plan:
PASS: During a factory install, verify that if the system type is
      duplex, the script is skipped, the correct message is logged,
      and the factory install is not blocked.
PASS: Verify that the factory install workflow is not modified for
      simplex.

Story: 2011455
Task: 52482

Change-Id: I5964af84de895f4fcbe83fb5173bb8461c715d5a
Signed-off-by: Enzo Candotti <Enzo.Candotti@windriver.com>