Add open-vcsr sample app for factory install
This commit adds two files to be executed in the setup stage of
factory install, aiming to install a customized application.
These files serve as examples of how to update the setup and config
content in factory-install.

The files added are:
- 20-vcsr-setup: Script added to the setup directory, executed after the
  system setup script. It performs pre-checks, installs the sample app,
  and verifies that the pods are Running.
- open-vcsr-example.yaml: YAML manifest applied by the script above,
  located in the config directory.

Test Plan:
PASS: Run a factory-install with these changes. Verify the new script
      runs successfully and, at the end of the stage, the new
      application is running as expected.

Story: 2011455
Task: 52572

Change-Id: I921c2dc6d32771ec282328fd1121b446e8c2136b
Signed-off-by: Enzo Candotti <Enzo.Candotti@windriver.com>