Run factory-install-backup service as root
This commit updates the privileges for the factory install backup
service, so that it runs as root. Currently, the service runs as the
sysadmin user, which causes issues when performing actions that
require elevated privileges.

Test Plan:
PASS: Run a factory install and verify that the backup service can
      perform the actions with root privileges, and the factory
      install finishes successfully.

Story: 2011455
Task: 52482

Change-Id: I9b530ef48dec4f7e126d3074315ffbef02601998
Signed-off-by: Enzo Candotti <Enzo.Candotti@windriver.com>