Openbao-monitor: Add store to k8s for init
Adding a store to k8s option for tokens and key shards in the monitor
config. This will only work if the useK8sConfigs option was enabled.

Test Plan:
PASS    The secrets are stored back after init in the k8s based openbao
server
PASS    The secrets are stored back after init in the host based openbao
server
PASS    The secrets are not erroneously stored in k8s secrets for
non-init operations in the k8s based openbao server

Story: 2011244
Task: 52697

Change-Id: Iac39a4391aae6713ae194850b37735d3efd385bd
Signed-off-by: Tae Park <tae.park@windriver.com>