Add certificate check and recovery on enroll init
Previously, if a subcloud remained staged long enough for certificates
that are not automatically renewed (and therefore must be manually
updated) to expire, the enrollment would fail late in the process.

This change updates the enroll-init-reconfigure script to check these
manual CA certificates such as etcd CA, k8s Root CA, Front Proxy CA,
and system-local-ca. If any are expired or close to expiration (within
3 days), the script fails and sends specific IPMI SEL events so each
failure is clear to the end user.

It also adds checks for leaf certificates. If they are expired or
near expiration, renewal steps are triggered. This stage also
includes specific IPMI SEL events.

On the success path, if there are no expired certificates, the time
added to the enrollment process is minimal, since all checks are
completed within a few seconds. If a renewal of leaf certificates is
required, the renewal process takes approximately 2m45s.

Test Plan:
01. PASS: Run subcloud enrollment and verify that the custom cloud-init
    complete IPMI SEL event is sent and that the enroll-init
    completes successfully.
02. PASS: Change the hardware clock of a subcloud so that, after
    powering it on, manual certificates are expired. Repeat for
    different dates to fail each manual certificate individually and
    in combination. Verify that each failure is detected, enroll-init
    fails, and the system controller logs show clear error messages.
    In particular, the scenarios tested are:
    A. <90 days: No certificates expired, so no recovery is needed,
    and enroll-init completes successfully.
    B. >90 and <360 days: Expired certificates are those
    automatically renewed by cert-manager. No recovery action is
    needed, and enroll-init completes successfully.
    C. >360 days: Kubernetes leaf certificates expired. Verify that
    expiration is detected, the renewal process runs successfully,
    and enroll-init completes correctly.
    D. >3060 days: k8s leaf and system-local-ca certificates expired.
    Verify that the leaf certificates renewal runs successfully, and
    the system-local-ca certificate expiration is detected. Therefore,
    enroll-init fails, and a specific message about the need for
    manual update is logged in the system controller.
    E. >3650 days: Kubernetes root CA, front-proxy CA, and etcd CA
    certificates expired. Verify that expirations are detected,
    enroll-init fails, and a specific message is logged in the
    system controller.
03. PASS: Test timeout scenarios by not sending IPMI events, verify
    that the playbook fails with appropriate timeout messages for each
    stage.

Partial-bug: 2119975

Change-Id: I0bcb6b1004466c8223ad831c0b8ea75debe385f0
Signed-off-by: Enzo Candotti <Enzo.Candotti@windriver.com>