Tox/Zuul: Change parent job to tox
Some STX Zuul jobs are using openstack jobs as parents.
The problem with that is that openstack has their own
upper-requirement files and job definitions and they aren't
always valid for the STX cases.

In this instance, openstack has changed their requirements
to versions that are not available in python3.9, which is
currently used by STX [1].

[1] https://review.opendev.org/c/openstack/requirements/+/958160

Test Plan:
pass: zuul

Partial-Bug: 2121603

Change-Id: I875d7571d424392a5c15e6927c78c30d72641485
Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>