Avoid bringing iface down before OAM reconfiguration
The oam modify command already updates the OAM configuration and
reconfigures the default route. For this to work, the IP address
must remain configured on the system to ensure the default gateway is
valid. In addition, if there is an interface change during enrollment,
the apply_network_config script adjusts the default route according to
the new interface.

This now fix the cases where ifdown removed the new subnet addresses,
leading to failures when replacing the default route since the
corresponding local route was no longer present.

Previously, the use of ifdown before OAM reconfiguration was introduced
to avoid cases where an IP address was applied while the same address
was already present, which caused operation failures. Now, ifup
includes a safeguard that skips the operation if the same address is
already configured on the interface, preventing issues caused by
duplicate IP assignments.

In addition, the removal of the /etc/network/interfaces.d/50-cloud-init
file was introduced after the enrollment operation is completed, since
keeping this file in the system would cause configuration
inconsistencies. For example, if there is a new OAM reconfiguration,
the address from 50-cloud-init would still persist after a lock/unlock
cycle, which would consequently lead to configuration inconsistencies.
In cases where OAM reconfiguration is not required, the file
50-cloud-init should already be removed at this stage, since the OAM
configuration puppet will not be executed and, consequently, neither
will apply_network_config.

Test Plan
=========

After enrollment, check that the 50-cloud-init file no longer exists
under /etc/network/interfaces.d. This verification must be performed
as part of all test cases.

[PASS] Execute subcloud enrollment process against a DX subcloud by
changing the bootstrap values IP to addresses from different subnets
than those configured in the subcloud and switching the bootstrap
interface from one VLAN X to another VLAN y. Verify that the default
route is updated and that the IP addresses on the VLAN are correct
after the unlock.

[PASS] Execute subcloud enrollment process against a DX subcloud by
changing the bootstrap values IP to addresses from different subnets
than those configured in the subcloud and switching the bootstrap
interface from VLAN to physical. Verify that the default route is
updated and that the IP addresses on the VLAN are correct after the
unlock.

[PASS] Execute subcloud enrollment process against a DX subcloud by
changing the bootstrap values IP to addresses from different subnets
than those configured in the subcloud. Verify that the default route
is updated and that IP addresses on the VLAN are correct before and
after the unlock.

[PASS] Execute subcloud enrollment process against a DX subcloud by
changing the bootstrap values IP to a different IP from the one
configured in the subcloud but within the same subnet (varying the
floating and node addresses). Verify that the default route, as well
as the floating and IP addresses on the VLAN, are correct before and
after the unlock.

[PASS] Execute subcloud enrollment process against a DX subcloud
by changing the bootstrap values IP to match the configured subcloud
IP and VLAN. Verify that the default route, as well as the IP addresses
on the VLAN, are correct before and after the unlock.

[PASS] Execute subcloud enrollment process against a SX subcloud by
changing the bootstrap values IP to addresses from different subnets
than those configured in the subcloud and switching the bootstrap
interface from VLAN to physical. Verify that the default route is
updated and that the IP addresses on the VLAN are correct after the
unlock.

[PASS] Execute subcloud enrollment process against a SX subcloud by
changing the bootstrap values IP to a different IP from the one
configured in the subcloud but within the same subnet (varying the
floating and node addresses). Verify that the default route, as well
as IP addresses on the VLAN, are correct before and after the unlock.

Closes-Bug: 2121954

Change-Id: I710638239ccaf36e0e498f5901ad13d9a7a37c15
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>